Removed DownloadHelper Updated Pdf viewwe Activity
